api/requirements.txt
Flask>=2.2.5
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
requests>=2.26.0
pymongo==3.11.4
Flask-PyMongo>=2.2.0
Werkzeug>=2.2.3
python-dotenv>=0.21.1
watchdog==3.0.0
gunicorn==21.2.0